refactor(providers): move local participant-config helpers beside their caller
They read the model stores, not settings state; settingsStore re-exports
for existing importers. Prepares the S2 participant-config seam, whose
local overrides must call these without a descriptor→settingsStore cycle.
